Issue fix
h3. Steps to reproduce

In Admin:
# Go to a CMS page. Add a "FAQ" element to the page
# Add a question
# Add an answer text to a WYSIWYG field
# Save the page
# Observe the FAQ content

(/) Expected result: 
the question text is visible. The answer text is visible by default (no need to go to the settings slideout panel of the FAQ item)
(x) Actual result: 
the answer text is not visible by default. It can be edited only in the settings slideout panel of the FAQ item